segment ectypes/server_py
