+++ ectypes/__init__.py
@@ -2,7 +2,7 @@ from .block_id import (
     BlockID,
 )
 
-from .block_desc import(
+from .block_desc import (
     BlockDesc,
 )
 
@@ -33,8 +33,14 @@ from .server import (
     idc_distance,
     make_serverrec,
     validate_idc,
+)
 
+from .server_id import (
     ServerID,
+)
+
+from .drive_id import (
+
     DriveID,
 )
 
@@ -71,6 +77,7 @@ __all__ = [
     "validate_idc",
 
     "ServerID",
+
     "DriveID",
 
     "Region",
